Workflow for TOFcalibCollector
It will reformat the calibration information coming from the matching
saving them to file, to be subsequently processed to extract the
time slewing (which does not seem appropriate for a data processor).

Some changes implemented also in the TimeSlotCalibration, when
the maxDelay is set to zero (--> we accept all data).

Just not to lose work :-)

compiling version

Changing philosophy.

Since we cannot send trees, the tree will have to be filled by the data processor
that will then also write the output file.
First version of the Spec also committed - but still using the tree. Now it is too
late, I will go to bed and take care another time.

Addng writer + other changes

Things compile, but were not yet tested.
Probably to change from std::vector to gsl::span.

Fixing issues + new features.

E.g.: using gsl::span for the output that goes in the tree.
Workflow not yet working, but everything compiles.

Fixes

Going back to std::vectors (tree cannot contain gsl::span)

Fixes to have the whole thing working. One change also to TimeSlotCalibration.h

The change in TimeSlotCalibration.h is needed because we want to update at every TF,
which means that the mMaxSlotsDelay should be 0.

fix for size of vector

I think this is correct, but I did not test it

move resize

fix for TOF channel offset calibration from reco flow

Several changes.

Main ones:

- bug fixes (seg fault due to missing initialization; merging function fixed)
- changes in the TimeSlotCalibration to not discard anything if maxDelay = 0
- force sending output at the end of the processing even if there is not
necessarily the requested statistics

Clang-format

code checker

clang-format

O2 code checker

+++ DataFormats/Detectors/TOF/include/DataFormatsTOF/CalibInfoTOF.h
@@ -36,7 +36,7 @@ class CalibInfoTOF
   void setDeltaTimePi(float time) { mDeltaTimePi = time; }
   float getDeltaTimePi() const { return mDeltaTimePi; }
 
-  void setTot(int tot) { mTot = tot; }
+  void setTot(float tot) { mTot = tot; }
   float getTot() const { return mTot; }
 
   void setFlags(int flags) { mFlags = flags; }
